Create controller, move logic into them, add controller to entry point.
